@@ -1,23 +1,24 @@
 name: native_dio_adapter
 description: A client for dio which makes use of cupertino_http and cronet_http to delegate HTTP requests to the native platform.
-version: 0.1.0
+version: 1.0.0+1
 homepage: https://github.com/cfug/dio
 repository: https://github.com/cfug/dio/blob/main/plugins/native_dio_adapter
 issue_tracker: https://github.com/cfug/dio/issues
 
 environment:
-  sdk: '>=2.19.0 <3.0.0'
-  flutter: ">=3.7.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.10.0"
 
 dependencies:
-  dio: ^5.0.0
-  cupertino_http: ^0.1.2
-  cronet_http: ^0.2.0
-  http: ^0.13.4
+  dio: ^5.2.0
+  cupertino_http: ^1.0.0
+  cronet_http: ^0.2.1
+  http: ^1.0.0
 
 dev_dependencies:
   lints: ^2.0.0
-  test: ^1.16.0
+  flutter_test:
+    sdk: flutter
 
 platforms:
   android:
