@@ -5,7 +5,7 @@ import 'dart:typed_data';
 
 import 'package:meta/meta.dart';
 import '../adapter.dart';
-import '../dio_error.dart';
+import '../dio_exception.dart';
 import '../headers.dart';
 import '../options.dart';
 
@@ -13,6 +13,8 @@ HttpClientAdapter createAdapter() => BrowserHttpClientAdapter();
 
 /// The default [HttpClientAdapter] for Web platforms.
 class BrowserHttpClientAdapter implements HttpClientAdapter {
+  BrowserHttpClientAdapter({this.withCredentials = false});
+
   /// These are aborted if the client is closed.
   @visibleForTesting
   final xhrs = <HttpRequest>{};
@@ -23,7 +25,7 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
   /// Defaults to `false`.
   ///
   /// You can also override this value in Options.extra['withCredentials'] for each request
-  bool withCredentials = false;
+  bool withCredentials;
 
   @override
   Future<ResponseBody> fetch(
@@ -49,10 +51,12 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
 
     final connectTimeout = options.connectTimeout;
     final receiveTimeout = options.receiveTimeout;
+    int xhrTimeout = 0;
     if (connectTimeout != null &&
         receiveTimeout != null &&
         receiveTimeout > Duration.zero) {
-      xhr.timeout = (connectTimeout + receiveTimeout).inMilliseconds;
+      xhrTimeout = (connectTimeout + receiveTimeout).inMilliseconds;
+      xhr.timeout = xhrTimeout;
     }
 
     final completer = Completer<ResponseBody>();
@@ -77,24 +81,16 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
       connectTimeoutTimer = Timer(
         connectionTimeout,
         () {
-          if (!completer.isCompleted) {
-            xhr.abort();
-            completer.completeError(
-              DioError.connectionTimeout(
-                requestOptions: options,
-                timeout: connectionTimeout,
-              ),
-              StackTrace.current,
-            );
-            return;
-          } else {
+          if (completer.isCompleted) {
             // connectTimeout is triggered after the fetch has been completed.
+            return;
           }
+
           xhr.abort();
           completer.completeError(
-            DioError.connectionTimeout(
+            DioException.connectionTimeout(
               requestOptions: options,
-              timeout: options.connectTimeout!,
+              timeout: connectionTimeout,
             ),
             StackTrace.current,
           );
@@ -120,7 +116,10 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
         if (duration > sendTimeout) {
           uploadStopwatch.stop();
           completer.completeError(
-            DioError.sendTimeout(timeout: sendTimeout, requestOptions: options),
+            DioException.sendTimeout(
+              timeout: sendTimeout,
+              requestOptions: options,
+            ),
             StackTrace.current,
           );
           xhr.abort();
@@ -150,7 +149,7 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
         if (duration > receiveTimeout) {
           downloadStopwatch.stop();
           completer.completeError(
-            DioError.receiveTimeout(
+            DioException.receiveTimeout(
               timeout: options.receiveTimeout!,
               requestOptions: options,
             ),
@@ -172,13 +171,28 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
       // specific information about the error itself.
       // See also: https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequestEventTarget/onerror
       completer.completeError(
-        DioError.connectionError(
+        DioException.connectionError(
           requestOptions: options,
           reason: 'The XMLHttpRequest onError callback was called. '
               'This typically indicates an error on the network layer.',
         ),
         StackTrace.current,
       );
+    });
+
+    xhr.onTimeout.first.then((_) {
+      if (connectTimeoutTimer != null) {
+        connectTimeoutTimer?.cancel();
+      }
+      if (!completer.isCompleted) {
+        completer.completeError(
+          DioException.receiveTimeout(
+            timeout: Duration(milliseconds: xhrTimeout),
+            requestOptions: options,
+          ),
+          StackTrace.current,
+        );
+      }
     });
 
     cancelFuture?.then((_) {
@@ -192,7 +206,7 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
         // or added xhr.onAbort like axios did https://github.com/axios/axios/blob/master/lib/adapters/xhr.js#L102-L111
         if (!completer.isCompleted) {
           completer.completeError(
-            DioError.requestCancelled(
+            DioException.requestCancelled(
               requestOptions: options,
               reason: 'The XMLHttpRequest was aborted.',
             ),
