@@ -1,3 +1,5 @@
+import 'package:meta/meta.dart';
+
 import 'adapter.dart';
 import 'cancel_token.dart';
 import 'headers.dart';
@@ -187,7 +189,7 @@ mixin OptionsMixin {
   late Map<String, dynamic> queryParameters;
 
   /// Timeout in milliseconds for opening url.
-  /// [Dio] will throw the [DioError] with [DioErrorType.connectionTimeout] type
+  /// [Dio] will throw the [DioException] with [DioExceptionType.connectionTimeout] type
   ///  when time out.
   Duration? get connectTimeout => _connectTimeout;
 
@@ -290,6 +292,7 @@ class Options {
     Options? options,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
+    StackTrace? sourceStackTrace,
   }) {
     final query = <String, dynamic>{};
     query.addAll(baseOpt.queryParameters);
@@ -315,6 +318,7 @@ class Options {
       baseUrl: baseOpt.baseUrl,
       path: path,
       data: data,
+      sourceStackTrace: sourceStackTrace ?? StackTrace.current,
       connectTimeout: baseOpt.connectTimeout,
       sendTimeout: sendTimeout ?? baseOpt.sendTimeout,
       receiveTimeout: receiveTimeout ?? baseOpt.receiveTimeout,
@@ -335,6 +339,7 @@ class Options {
       cancelToken: cancelToken,
       contentType: contentType ?? this.contentType ?? baseOpt.contentType,
     );
+    requestOptions.cancelToken?.requestOptions = requestOptions;
     return requestOptions;
   }
 
@@ -349,7 +354,7 @@ class Options {
   Map<String, dynamic>? headers;
 
   /// Timeout in milliseconds for sending data.
-  /// [Dio] will throw the [DioError] with [DioErrorType.sendTimeout] type
+  /// [Dio] will throw the [DioException] with [DioExceptionType.sendTimeout] type
   ///  when time out.
   Duration? get sendTimeout => _sendTimeout;
 
@@ -471,6 +476,7 @@ class RequestOptions extends _RequestConfig with OptionsMixin {
     ResponseDecoder? responseDecoder,
     ListFormat? listFormat,
     bool? setRequestContentTypeWhenNoPayload,
+    StackTrace? sourceStackTrace,
   })  : assert(connectTimeout == null || !connectTimeout.isNegative),
         super(
           method: method,
@@ -489,6 +495,7 @@ class RequestOptions extends _RequestConfig with OptionsMixin {
           responseDecoder: responseDecoder,
           listFormat: listFormat,
         ) {
+    this.sourceStackTrace = sourceStackTrace ?? StackTrace.current;
     this.queryParameters = queryParameters ?? {};
     this.baseUrl = baseUrl ?? '';
     this.connectTimeout = connectTimeout;
@@ -555,6 +562,7 @@ class RequestOptions extends _RequestConfig with OptionsMixin {
       requestEncoder: requestEncoder ?? this.requestEncoder,
       responseDecoder: responseDecoder ?? this.responseDecoder,
       listFormat: listFormat ?? this.listFormat,
+      sourceStackTrace: sourceStackTrace,
     );
 
     if (contentType != null) {
@@ -566,6 +574,14 @@ class RequestOptions extends _RequestConfig with OptionsMixin {
 
     return ro;
   }
+
+  /// The source [StackTrace] which should always point to
+  /// the invocation of [DioMixin.request] or if not provided,
+  /// to the construction of the [RequestOptions] instance.
+  /// In both instances the source context should still be
+  /// available before it is lost due to asynchronous operations.
+  @internal
+  StackTrace? sourceStackTrace;
 
   /// generate uri
   Uri get uri {
@@ -675,7 +691,7 @@ class _RequestConfig {
   }
 
   /// Timeout in milliseconds for sending data.
-  /// [Dio] will throw the [DioError] with [DioErrorType.sendTimeout] type
+  /// [Dio] will throw the [DioException] with [DioExceptionType.sendTimeout] type
   ///  when time out.
   ///
   /// `null` meanings no timeout limit.
