@@ -25,7 +25,7 @@ class CacheInterceptor extends Interceptor {
   }
 
   @override
-  void onError(DioError err, ErrorInterceptorHandler handler) {
+  void onError(DioException err, ErrorInterceptorHandler handler) {
     print('onError: $err');
     super.onError(err, handler);
   }
