@@ -6,7 +6,7 @@ library dio;
 export 'src/adapter.dart';
 export 'src/cancel_token.dart';
 export 'src/dio.dart';
-export 'src/dio_error.dart';
+export 'src/dio_exception.dart';
 export 'src/dio_mixin.dart' hide InterceptorState, InterceptorResultType;
 export 'src/form_data.dart';
 export 'src/headers.dart';
