@@ -1,4 +1,4 @@
-@TestOn('chrome')
+@TestOn('browser')
 import 'dart:typed_data';
 
 import 'package:dio/browser.dart';
@@ -7,11 +7,12 @@ import 'package:test/test.dart';
 
 void main() {
   test('with credentials', () async {
-    final browserAdapter = BrowserHttpClientAdapter();
-    browserAdapter.withCredentials = true;
+    final browserAdapter = BrowserHttpClientAdapter(withCredentials: true);
     final opts = RequestOptions();
-    final testStream =
-        Stream<Uint8List>.periodic(Duration(seconds: 1), (x) => Uint8List(x));
+    final testStream = Stream<Uint8List>.periodic(
+      Duration(seconds: 1),
+      (x) => Uint8List(x),
+    );
     final cancelFuture = opts.cancelToken?.whenCancel;
 
     browserAdapter.fetch(opts, testStream, cancelFuture);
