@@ -8,18 +8,20 @@ void main() async {
   dio.options
     ..headers['user-agent'] = 'xxx'
     ..contentType = 'text';
-  // dio.options.connectTimeout = 2000;
-  // More about HttpClient proxy topic please refer to Dart SDK doc.
-  dio.httpClientAdapter = IOHttpClientAdapter()
-    ..onHttpClientCreate = (HttpClient client) {
+  dio.httpClientAdapter = IOHttpClientAdapter(
+    createHttpClient: () {
+      final client = HttpClient();
       client.findProxy = (uri) {
-        //proxy all request to localhost:8888
+        // Proxy all request to localhost:8888.
+        // Be aware, the proxy should went through you running device,
+        // not the host platform.
         return 'PROXY localhost:8888';
       };
       client.badCertificateCallback =
           (X509Certificate cert, String host, int port) => true;
       return client;
-    };
+    },
+  );
 
   Response<String> response;
   response = await dio.get('https://www.baidu.com');
