@@ -6,15 +6,13 @@ void main() async {
     ..options.baseUrl = 'https://pub.dev'
     ..interceptors.add(LogInterceptor())
     ..httpClientAdapter = Http2Adapter(
-      ConnectionManager(
-        idleTimeout: Duration(seconds: 10),
-      ),
+      ConnectionManager(idleTimeout: Duration(seconds: 10)),
     );
 
   Response<String> response;
-
   response = await dio.get('/?xx=6');
-  print(response.data?.length);
-  print(response.redirects.length);
+  for (final e in response.redirects) {
+    print('redirect: ${e.statusCode} ${e.location}');
+  }
   print(response.data);
 }
