@@ -26,12 +26,25 @@ void main() {
 
   test('request with payload', () async {
     final dio = Dio()
-      ..options.baseUrl = 'https://httpbin.org/'
+      ..options.baseUrl = 'https://httpbun.com/'
       ..httpClientAdapter = Http2Adapter(
         ConnectionManager(
           idleTimeout: Duration(milliseconds: 10),
         ),
       );
+
+    final res = await dio.post('post', data: 'TEST');
+    expect(res.data.toString(), contains('TEST'));
+  });
+
+  test('request with payload via proxy', () async {
+    final dio = Dio()
+      ..options.baseUrl = 'https://httpbun.com/'
+      ..httpClientAdapter = Http2Adapter(ConnectionManager(
+        idleTimeout: Duration(milliseconds: 10),
+        onClientCreate: (uri, settings) =>
+            settings.proxy = Uri.parse('http://localhost:3128'),
+      ));
 
     final res = await dio.post('post', data: 'TEST');
     expect(res.data.toString(), contains('TEST'));
